+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = "eBird API Data"
 copyright = "2025, Stuart MacKay"
 author = "Stuart MacKay"
-release = "0.1.1"
+release = "0.2.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration